requirements.txt
numpy~=1.18.1
py4j~=0.10.9
scipy~=1.4.1
zlib~=1.2.11
setuptools~=41.2.0
pymongo~=3.11.4
tweepy~=3.10.0
textblob~=0.15.3
kafka-python==0.9.4
oauthlib==1.0.3
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy==3.3.0
configparser==3.3.0r2
matplotlib==1.5.0
boto3~=1.17.93
botocore~=1.20.93